backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
Pillow
pdf2image
transformers
torch
opencv-python
numpy
sentencepiece
protobuf
redis
rq
groq
python-dotenv
pytesseract
google-genai
